Fix tests running infinitely
The @shelf/jest-mongodb required configuration of 'preset': ... broke
the existing RecurringJob tests illuminating the fact that the jobs
are never canceled. Not sure how to have only a preset for a subset of
jobs, but was able to cancel all jobs after each test. Also added
coverage flag to the logging jest call.