Add basic message response logging functionality
Adds a messageResponse model that saves the number the response is from,
the message, and the message it is responding too. The message it is
responding too is the last message sent to the number that is responding.
This does not handle multiple responses to the same message, or multiple
people messaging the same number and them responding to only one.
(Adding these as issues).
If it cannot find a message sent to the number responding, no message
is saved.