Fix message response attribution (#66)
Fixes #65. The sort on the sms-response was based on an attribute that didn't exist
so it wasn't doing anything. It has been fixed, and the same sorting
has been added to both messages-sent and messages-received endpoints
so they are sorted from newest to oldest message. Message response
attribution is now correct and the dashboard correctly calculates
the number of responses.